Add infinity constant
In OMSI Script dividing by 0 evaluates to 0 rather than infinity,
undefined or an exception. This constant can be used in such situations
where a value of infinity is desirable. This is achieved by trying to
store a value of 4x10^38 into a 32-bit float constant that can only
handle up to ~3.4x10^38.